+++ src/Components/Modules/FrameTime.hpp
@@ -11,10 +11,10 @@ namespace Components
 		static void SVFrameWaitStub();
 		static void SVFrameWaitFunc();
 
-		static void NetSleep(int msec);
+		static void NetSleep(int mSec);
 
 		static int ComTimeVal(int minMsec);
-		static uint32_t ComFrameWait(int minMsec);
+		static int ComFrameWait(int minMsec);
 		static void ComFrameWaitStub();
 	};
 }
