+++ src/Game/Functions.cpp
@@ -23,23 +23,6 @@ namespace Game
 	Cmd_AddCommand_t Cmd_AddCommand = Cmd_AddCommand_t(0x470090);
 	Cmd_AddServerCommand_t Cmd_AddServerCommand = Cmd_AddServerCommand_t(0x4DCE00);
 	Cmd_ExecuteSingleCommand_t Cmd_ExecuteSingleCommand = Cmd_ExecuteSingleCommand_t(0x609540);
-	Com_ClientPacketEvent_t Com_ClientPacketEvent = Com_ClientPacketEvent_t(0x49F0B0);
-
-	Com_Error_t Com_Error = Com_Error_t(0x4B22D0);
-	Com_Printf_t Com_Printf = Com_Printf_t(0x402500);
-	Com_PrintError_t Com_PrintError = Com_PrintError_t(0x4F8C70);
-	Com_PrintWarning_t Com_PrintWarning = Com_PrintWarning_t(0x4E0200);
-	Com_PrintMessage_t Com_PrintMessage = Com_PrintMessage_t(0x4AA830);
-	Com_EndParseSession_t Com_EndParseSession = Com_EndParseSession_t(0x4B80B0);
-	Com_BeginParseSession_t Com_BeginParseSession = Com_BeginParseSession_t(0x4AAB80);
-	Com_ParseOnLine_t Com_ParseOnLine = Com_ParseOnLine_t(0x4C0350);
-	Com_SkipRestOfLine_t Com_SkipRestOfLine = Com_SkipRestOfLine_t(0x4B8300);
-	Com_SetSpaceDelimited_t Com_SetSpaceDelimited = Com_SetSpaceDelimited_t(0x4FC710);
-	Com_Parse_t Com_Parse = Com_Parse_t(0x474D60);
-	Com_MatchToken_t Com_MatchToken = Com_MatchToken_t(0x447130);
-	Com_SetSlowMotion_t Com_SetSlowMotion = Com_SetSlowMotion_t(0x446E20);
-	Com_Quitf_t Com_Quit_f = Com_Quitf_t(0x4D4000);
-	Com_OpenLogFile_t Com_OpenLogFile = Com_OpenLogFile_t(0x60A8D0);
 
 	Con_DrawMiniConsole_t Con_DrawMiniConsole = Con_DrawMiniConsole_t(0x464F30);
 	Con_DrawSolidConsole_t Con_DrawSolidConsole = Con_DrawSolidConsole_t(0x5A5040);
@@ -317,8 +300,6 @@ namespace Game
 
 	SOCKET* ip_socket = reinterpret_cast<SOCKET*>(0x64A3008);
 
-	uint32_t* com_frameTime = reinterpret_cast<uint32_t*>(0x1AD8F3C);
-
 	SafeArea* safeArea = reinterpret_cast<SafeArea*>(0xA15F3C);
 
 	SpawnVar* spawnVars = reinterpret_cast<SpawnVar*>(0x1A83DE8);
@@ -402,10 +383,6 @@ namespace Game
 	unsigned long* _tls_index = reinterpret_cast<unsigned long*>(0x66D94A8);
 
 	int* cls_uiStarted = reinterpret_cast<int*>(0xA7FFA0);
-
-	int* com_fixedConsolePosition = reinterpret_cast<int*>(0x1AD8EC8);
-
-	int* com_errorPrintsCount = reinterpret_cast<int*>(0x1AD7910);
 
 	unsigned int* playerCardUIStringIndex = reinterpret_cast<unsigned int*>(0x62CD7A8);
 	char (*playerCardUIStringBuf)[PLAYER_CARD_UI_STRING_COUNT][38] = reinterpret_cast<char(*)[PLAYER_CARD_UI_STRING_COUNT][38]>(0x62CB4F8);
@@ -520,38 +497,6 @@ namespace Game
 			{
 				(*storeHere)->Unlock();
 			}
-		}
-	}
-
-	char* Com_GetParseThreadInfo()
-	{
-		if (Sys_IsMainThread())
-		{
-			return reinterpret_cast<char*>(0x6466628);
-		}
-		if (Sys_IsRenderThread())
-		{
-			return reinterpret_cast<char*>(0x646AC34);
-		}
-		if (Sys_IsServerThread())
-		{
-			return reinterpret_cast<char*>(0x646F240);
-		}
-		if (Sys_IsDatabaseThread())
-		{
-			return reinterpret_cast<char*>(0x647384C);
-		}
-
-		return nullptr;
-	}
-
-	void Com_SetParseNegativeNumbers(int parse)
-	{
-		char* g_parse = Com_GetParseThreadInfo();
-
-		if (g_parse)
-		{
-			g_parse[1056 * *(reinterpret_cast<DWORD*>(g_parse) + 4224) + 1032] = parse != 0;
 		}
 	}
