+++ src/Game/Functions.hpp
@@ -54,54 +54,6 @@ namespace Game
 	typedef void(*Cmd_ExecuteSingleCommand_t)(int localClientNum, int controllerIndex, const char* cmd);
 	extern Cmd_ExecuteSingleCommand_t Cmd_ExecuteSingleCommand;
 
-	typedef void(*Com_ClientPacketEvent_t)();
-	extern Com_ClientPacketEvent_t Com_ClientPacketEvent;
-
-	typedef void(*Com_Error_t)(errorParm_t type, const char* message, ...);
-	extern Com_Error_t Com_Error;
-
-	typedef void(*Com_Printf_t)(int channel, const char* fmt, ...);
-	extern Com_Printf_t Com_Printf;
-
-	typedef void(*Com_PrintError_t)(int channel, const char* fmt, ...);
-	extern Com_PrintError_t Com_PrintError;
-
-	typedef void(*Com_PrintWarning_t)(int channel, const char* fmt, ...);
-	extern Com_PrintWarning_t  Com_PrintWarning;
-
-	typedef void(*Com_PrintMessage_t)(int channel, const char* msg, int error);
-	extern Com_PrintMessage_t Com_PrintMessage;
-
-	typedef void(*Com_EndParseSession_t)();
-	extern Com_EndParseSession_t Com_EndParseSession;
-
-	typedef void(*Com_BeginParseSession_t)(const char* filename);
-	extern Com_BeginParseSession_t Com_BeginParseSession;
-
-	typedef char*(*Com_ParseOnLine_t)(const char** data_p);
-	extern Com_ParseOnLine_t Com_ParseOnLine;
-
-	typedef void(*Com_SkipRestOfLine_t)(const char** data);
-	extern Com_SkipRestOfLine_t Com_SkipRestOfLine;
-
-	typedef void(*Com_SetSpaceDelimited_t)(int);
-	extern Com_SetSpaceDelimited_t Com_SetSpaceDelimited;
-
-	typedef char*(*Com_Parse_t)(const char** data_p);
-	extern Com_Parse_t Com_Parse;
-
-	typedef bool (*Com_MatchToken_t)(const char **data_p, const char* token, int size);
-	extern Com_MatchToken_t Com_MatchToken;
-
-	typedef void(*Com_SetSlowMotion_t)(float start, float end, int duration);
-	extern Com_SetSlowMotion_t Com_SetSlowMotion;
-
-	typedef void(*Com_Quitf_t)();
-	extern Com_Quitf_t Com_Quit_f;
-
-	typedef void(*Com_OpenLogFile_t)();
-	extern Com_OpenLogFile_t Com_OpenLogFile;
-
 	typedef char* (*Con_DrawMiniConsole_t)(int localClientNum, int xPos, int yPos, float alpha);
 	extern Con_DrawMiniConsole_t Con_DrawMiniConsole;
 
@@ -666,8 +618,6 @@ namespace Game
 	extern netadr_t* connectedHost;
 	extern SOCKET* ip_socket;
 
-	extern uint32_t* com_frameTime;
-
 	extern SafeArea* safeArea;
 
 	extern SpawnVar* spawnVars;
@@ -756,10 +706,6 @@ namespace Game
 
 	extern int* cls_uiStarted;
 
-	extern int* com_fixedConsolePosition;
-
-	extern int* com_errorPrintsCount;
-
 	constexpr std::size_t PLAYER_CARD_UI_STRING_COUNT = 18;
 	extern unsigned int* playerCardUIStringIndex;
 	extern char (*playerCardUIStringBuf)[PLAYER_CARD_UI_STRING_COUNT][38];
@@ -805,9 +751,6 @@ namespace Game
 
 	void Load_IndexBuffer(void* data, IDirect3DIndexBuffer9** storeHere, int count);
 	void Load_VertexBuffer(void* data, IDirect3DVertexBuffer9** where, int len);
-
-	char* Com_GetParseThreadInfo();
-	void Com_SetParseNegativeNumbers(int parse);
 
 	void Image_Setup(GfxImage* image, unsigned int width, unsigned int height, unsigned int depth, unsigned int flags, _D3DFORMAT format);
